Pull out all of the common functions between gravimetric and photometric
